# docs/api/api.yaml
openapi: "3.0.0"
info:
  title: "Wutsi Catalog API"
  description: API for managing product catalog
  version: "1.0.0"

servers:
  - url: https://wutsi-catalog-server-test.herokuapp.com
    description: Sandbox
  - url: https://wutsi-catalog-server-prod.herokuapp.com
    description: Production

paths:
  /v1/categories/search:
    post:
      operationId: "search-categories"
      description: Search product categories
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchCategoryRequest'
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchCategoryResponse'
        403:
          description: The caller is not authorized to call this endpoint
        409:
          description: Error
      tags:
        - Category
      security:
        - api_key: [ 'catalog-manage' ]

  /v1/categories/{id}:
    get:
      operationId: "get-category"
      description: Return information about a category
      parameters:
        - in: path
          name: id
          description: ID of the category
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetCategoryResponse'
        403:
          description: The caller is not authorized to call this endpoint
        404:
          description: The category not found
      tags:
        - Product
      security:
        - api_key: [ 'catalog-read' ]



  /v1/sections:
    get:
      operationId: "list-sections"
      description: Return all the sections of a catalog
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListSectionResponse'
        403:
          description: The caller is not authorized to call this endpoint
      tags:
        - Section
      security:
        - api_key: [ 'catalog-read' ]

    post:
      operationId: "create-section"
      description: Create a new section
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSectionRequest'
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateSectionResponse'
        403:
          description: The caller is not authorized to call this endpoint
        409:
          description: Error
      tags:
        - Product
      security:
        - api_key: [ 'catalog-manage' ]

  /v1/sections/{id}:
    get:
      operationId: "get-section"
      description: Return a section
      parameters:
        - in: path
          name: id
          description: ID of the section
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetSectionResponse'
        403:
          description: The caller is not authorized to call this endpoint
      tags:
        - Section
      security:
        - api_key: [ 'catalog-read' ]

    post:
      operationId: "update-section"
      description: Update a section
      parameters:
        - in: path
          name: id
          description: ID of the section
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSectionRequest'
      responses:
        200:
          description: Success
        403:
          description: The caller is not authorized to call this endpoint
        409:
          description: Error
      tags:
        - Product
      security:
        - api_key: [ 'catalog-manage' ]

    delete:
      operationId: "delete-section"
      description: Delete a section
      parameters:
        - in: path
          name: id
          description: ID of the section
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Success
        403:
          description: The caller is not authorized to call this endpoint
      tags:
        - Section
      security:
        - api_key: [ 'catalog-manage' ]

  /v1/sections/{id}/products/{product-id}:
    post:
      operationId: "add-to-section"
      description: Add a product into a section
      parameters:
        - in: path
          name: id
          description: ID of the section
          required: true
          schema:
            type: integer
            format: int64
        - in: path
          name: product-id
          description: ID of the product
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Success
        403:
          description: The caller is not authorized to call this endpoint
      tags:
        - Section
      security:
        - api_key: [ 'catalog-manage' ]

    delete:
      operationId: "remove-from-section"
      description: Remove a product from a section
      parameters:
        - in: path
          name: id
          description: ID of the section
          required: true
          schema:
            type: integer
            format: int64
        - in: path
          name: product-id
          description: ID of the product
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Success
        403:
          description: The caller is not authorized to call this endpoint
      tags:
        - Section
      security:
        - api_key: [ 'catalog-manage' ]



  /v1/products:
    post:
      operationId: "create-product"
      description: Create a new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductRequest'
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateProductResponse'
        403:
          description: The caller is not authorized to call this endpoint
        409:
          description: Error
      tags:
        - Product
      security:
        - api_key: [ 'catalog-manage' ]

  /v1/products/search:
    post:
      operationId: "search-products"
      description: Search Product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchProductRequest'
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchProductResponse'
        403:
          description: The caller is not authorized to call this endpoint
      tags:
        - Product
      security:
        - api_key: [ 'catalog-read' ]

  /v1/products/{id}:
    get:
      operationId: "get-product"
      description: Return information about an product
      parameters:
        - in: path
          name: id
          description: ID of the product
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetProductResponse'
        403:
          description: The caller is not authorized to call this endpoint
        404:
          description: The product not found
      tags:
        - Product
      security:
        - api_key: [ 'catalog-read' ]

    delete:
      operationId: "delete-product"
      description: Delete an product
      parameters:
        - in: path
          name: id
          description: ID of the product to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Success
        403:
          description: The caller is not the owner of the product
      tags:
        - Product
      security:
        - api_key: [ 'catalog-manage' ]

  /v1/products/{id}/attributes/{name}:
    post:
      operationId: "update-product-attribute"
      description: Update an product attribute
      parameters:
        - in: path
          name: id
          description: ID of the product
          required: true
          schema:
            type: integer
            format: int64
        - in: path
          name: name
          description: Name of the attribute to update
          required: true
          schema:
            type: string
            enum:
              - title
              - summary
              - description
              - price
              - comparable-price
              - thumbnail-id
              - visible
              - sub-category-id
              - quantity
              - max-order
              - type
              - numeric-file-url
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProductAttributeRequest'
      responses:
        200:
          description: Success
        403:
          description: The caller is not the owner of the product
        404:
          description: The product not found
      tags:
        - Product
      security:
        - api_key: [ 'catalog-manage' ]

  /v1/products/{id}/pictures:
    post:
      operationId: "add-picture"
      description: Add a picture to a product
      parameters:
        - in: path
          name: id
          description: ID of the product
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddPictureRequest'
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddPictureResponse'
        403:
          description: The caller is not the owner of the product
        404:
          description: The product not found
      tags:
        - Product
      security:
        - api_key: [ 'catalog-manage' ]

  /v1/products/pictures/{id}:
    delete:
      operationId: "delete-picture"
      description: Delete a picture
      parameters:
        - in: path
          name: id
          description: ID of the picture
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Success
      tags:
        - Product
      security:
        - api_key: [ 'catalog-manage' ]


  /v1/reservations:
    post:
      operationId: "create-reservation"
      description: Reserve product for purchase
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReservationRequest'
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateReservationResponse'
      tags:
        - Reservaion
      security:
        - api_key: [ 'catalog-manage' ]

  /v1/reservations/{id}:
    delete:
      operationId: "cancel-reservation"
      description: Cancel a reservation
      parameters:
        - in: path
          name: id
          description: ID of the reservation
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Success
      tags:
        - Reservaion
      security:
        - api_key: [ 'catalog-manage' ]

components:
  securitySchemes:
    api_key:
      type: apiKey
      in: header
      name: Authorization

  schemas:
    CreateProductRequest:
      type: object
      properties:
        title:
          type: string
          description: Product title
          maxLength: 100
        summary:
          type: string
          description: Product summary
          maxLength: 160
          nullable: true
        description:
          type: string
          description: Product long description
          nullable: true
        price:
          type: number
          format: double
          description: Price of the product
          nullable: true
        comparablePrice:
          type: number
          format: double
          description: Comparable Price
          nullable: true
        categoryId:
          type: integer
          format: int64
          description: ID of the category
        subCategoryId:
          type: integer
          format: int64
          description: ID of the sub-category
        quantity:
          type: integer
          description: Quantity in stock
          minimum: 0
        maxOrder:
          type: integer
          description: Max items that a customer can order
          nullable: true
        type:
          type: string
          description: Type of product
          enum:
            - PHYSISCAL
            - NUMERIC
      required:
        - title
        - type

    CreateProductResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID of the product created

    GetProductResponse:
      type: object
      properties:
        product:
          type: object
          $ref: "#/components/schemas/Product"

    SearchProductRequest:
      type: object
      properties:
        productIds:
          type: array
          items:
            type: integer
            format: int64
        categoryIds:
          type: array
          items:
            type: integer
            format: int64
        sectionId:
          type: integer
          format: int64
          nullable: true
        visible:
          type: boolean
          nullable: true
        accountId:
          type: integer
          format: int64
          nullable: true
        limit:
          type: integer
          default: 30
        offset:
          type: integer
          default: 0

    SearchProductResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/ProductSummary"

    PictureSummary:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier of the product
        url:
          type: string
          format: url
          description: URL of picture

    ProductSummary:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier of the product
        accountId:
          type: integer
          format: int64
          description: ID of the merchant account
        thumbnail:
          type: object
          $ref: "#/components/schemas/PictureSummary"
          nullable: true
        title:
          type: string
          description: Product title
        summary:
          type: string
          description: Product summary
          nullable: true
        price:
          type: number
          format: double
          description: Price of the product
          nullable: true
        comparablePrice:
          type: number
          format: double
          description: Comparable Price
          nullable: true
        categoryId:
          type: integer
          format: int64
          description: ID of the category
        subCategoryId:
          type: integer
          format: int64
          description: ID of the sub-category
        quantity:
          type: integer
          description: Quantity in stock
        maxOrder:
          type: integer
          description: Max items that a customer can order
          nullable: true
        type:
          type: string
          description: Type of product
          enum:
            - PHYSISCAL
            - NUMERIC
        currency:
          type: string
          description: Product's country
        created:
          type: string
          format: date-time
          description: Creation Date/Time
        updated:
          type: string
          format: date-time
          description: Last modification Date/Time

    Product:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier of the product
        accountId:
          type: integer
          format: int64
          description: ID of the merchant account
        thumbnail:
          type: object
          $ref: "#/components/schemas/PictureSummary"
          nullable: true
        pictures:
          type: array
          items:
            $ref: "#/components/schemas/PictureSummary"
        category:
          type: object
          $ref: "#/components/schemas/CategorySummary"
        subCategory:
          type: object
          $ref: "#/components/schemas/CategorySummary"
        title:
          type: string
          description: Product title
        summary:
          type: string
          description: Product summary
          nullable: true
        description:
          type: string
          description: Product long description
          nullable: true
        price:
          type: number
          format: double
          description: Price of the product
          nullable: true
        comparablePrice:
          type: number
          format: double
          description: Comparable Price
          nullable: true
        currency:
          type: string
          description: Product's currency
        visible:
          type: boolean
          description: is the product visible
        categoryId:
          type: integer
          format: int64
          description: ID of the category
        subCategoryId:
          type: integer
          format: int64
          description: ID of the sub-category
        quantity:
          type: integer
          description: Quantity in stock
        maxOrder:
          type: integer
          description: Max items that a customer can order
          nullable: true
        type:
          type: string
          description: Type of product
          enum:
            - PHYSISCAL
            - NUMERIC
        created:
          type: string
          format: date-time
          description: Creation Date/Time
        updated:
          type: string
          format: date-time
          description: Last modification Date/Time
        numericFileUrl:
          type: string
          format: url
          description: For numeric product, this represent the URL of the associated file
          nullable: true
        sections:
          type: array
          items:
            $ref: "#/components/schemas/SectionSummary"

    AddPictureRequest:
      type: object
      properties:
        url:
          type: string
          format: url
          description: URL of the picture
      required:
        - url

    AddPictureResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID of the picture

    UpdateProductAttributeRequest:
      type: object
      properties:
        value:
          type: string
          description: Value of the
          nullable: true


    Category:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID of the category
        title:
          type: string
        parentId:
          type: integer
          format: int64
          description: ID of the parent category
          nullable: true

    GetCategoryResponse:
      type: object
      properties:
        category:
          type: object
          $ref: "#/components/schemas/Category"

    CategorySummary:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID of the category
        title:
          type: string
        parentId:
          type: integer
          format: int64
          description: ID of the parent category
          nullable: true

    SearchCategoryRequest:
      type: object
      properties:
        parentId:
          type: integer
          format: int64
          nullable: true

    SearchCategoryResponse:
      type: object
      properties:
        categories:
          type: array
          items:
            $ref: "#/components/schemas/CategorySummary"


    ReservationProduct:
      type: object
      properties:
        productId:
          type: integer
          format: int64
          description: ID of the product
        quantity:
          type: integer
          description: Quantity of product to reserve

    CreateReservationRequest:
      type: object
      properties:
        orderId:
          type: string
          format: uuid
          description: ID of the Order
        products:
          type: array
          items:
            $ref: "#/components/schemas/ReservationProduct"
          description: products to reserve

    CreateReservationResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID of the reservation

    SectionSummary:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID of the section
        title:
          type: string
          description: Title of the section
        sortOrder:
          type: integer
          description: sort order
        productCount:
          type: integer
          description: Number of products in this section

    ListSectionResponse:
      type: object
      properties:
        sections:
          type: array
          items:
            $ref: "#/components/schemas/SectionSummary"
          description: List of sections

    CreateSectionRequest:
      type: object
      properties:
        title:
          type: string
          description: Title of the section
      required:
        - title

    CreateSectionResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID of the section

    Section:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID of the section
        title:
          type: string
          description: Title of the section
        sortOrder:
          type: integer
          description: sort order
        productCount:
          type: integer
          description: Number of products in this section

    GetSectionResponse:
      type: object
      properties:
        section:
          type: object
          $ref: "#/components/schemas/Section"

    UpdateSectionRequest:
      type: object
      properties:
        title:
          type: string
          description: Title of the section
        sortOrder:
          type: integer
          description: sort order
      required:
        - title